Added more control buttons and sections
Added two sections for cards. One for the dealers cards
and another one for the player cards.

I also divided all the buttons is sections so they can be styled
according to my wireframe. I found making the sections on <div>s made it
easier to move them as block elements.